MyApps-13: Move get selected book UseCase to submodules.
