+++ app/src/main/java/xyz/torquato/myapps/di/domain/DomainWebModule.kt
@@ -4,11 +4,9 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import xyz.torquato.myapps.domain.web.GetSelectedBookUseCase
 import xyz.torquato.myapps.domain.web.content.GetBooksUseCase
 import xyz.torquato.myapps.domain.web.content.GetCacheBooksUseCase
 import xyz.torquato.myaps.domain.api.pagging.IPagingRepository
-import xyz.torquato.myaps.domain.api.web.ILocalWebRepository
 import xyz.torquato.myaps.domain.api.web.IQueryRepository
 import xyz.torquato.myaps.domain.api.web.IWebRepository
 import xyz.torquato.myaps.domain.api.web.cache.IWebCacheRepository
@@ -40,13 +38,7 @@ object DomainWebModule {
         webCacheRepository
     )
 
-    @Singleton
-    @Provides
-    fun provideGetSelectedBookIdUseCase(
-        repository: ILocalWebRepository
-    ): GetSelectedBookUseCase = GetSelectedBookUseCase(
-        repository
-    )
+
 
 
 }
